Move types to more appropriate namespaces
